+++ tests/RedisConnectionTest.php
@@ -1,12 +1,23 @@
 <?php
 
 namespace yiiunit\extensions\redis;
+use Yii;
+use yii\helpers\ArrayHelper;
+use yii\log\Logger;
+use yii\redis\Connection;
+use yii\redis\SocketException;
 
 /**
  * @group redis
  */
 class ConnectionTest extends TestCase
 {
+    protected function tearDown()
+    {
+        $this->getConnection(false)->configSet('timeout', 0);
+        parent::tearDown();
+    }
+
     /**
      * test connection to redis and selection of db
      */
@@ -87,6 +98,77 @@ class ConnectionTest extends TestCase
         $this->assertTrue($db2->ping());
     }
 
+    public function testConnectionTimeout()
+    {
+        $db = $this->getConnection(false);
+        $db->configSet('timeout', 1);
+        $this->assertTrue($db->ping());
+        sleep(1);
+        $this->assertTrue($db->ping());
+        sleep(2);
+        if (method_exists($this, 'setExpectedException')) {
+            $this->setExpectedException('\yii\redis\SocketException');
+        } else {
+            $this->expectException('\yii\redis\SocketException');
+        }
+        $this->assertTrue($db->ping());
+    }
+
+    public function testConnectionTimeoutRetry()
+    {
+        $logger = new Logger();
+        Yii::setLogger($logger);
+
+        $db = $this->getConnection(false);
+        $db->retries = 1;
+        $db->configSet('timeout', 1);
+        $this->assertCount(3, $logger->messages, 'log of connection and init commands.');
+
+        $this->assertTrue($db->ping());
+        $this->assertCount(4, $logger->messages, 'log +1 ping command.');
+        usleep(500000); // 500ms
+
+        $this->assertTrue($db->ping());
+        $this->assertCount(5, $logger->messages, 'log +1 ping command.');
+        sleep(2);
+
+        // reconnect should happen here
+
+        $this->assertTrue($db->ping());
+        $this->assertCount(11, $logger->messages, 'log +1 ping command, and reconnection.'
+            . print_r(array_map(function($s) { return (string) $s; }, ArrayHelper::getColumn($logger->messages, 0)), true));
+    }
+
+    /**
+     * Retry connecting 2 times
+     */
+    public function testConnectionTimeoutRetryCount()
+    {
+        $logger = new Logger();
+        Yii::setLogger($logger);
+
+        $db = $this->getConnection(false);
+        $db->retries = 2;
+        $db->configSet('timeout', 1);
+        $db->on(Connection::EVENT_AFTER_OPEN, function() {
+            // sleep 2 seconds after connect to make every command time out
+            sleep(2);
+        });
+        $this->assertCount(3, $logger->messages, 'log of connection and init commands.');
+
+        $exception = false;
+        try {
+            // should try to reconnect 2 times, before finally failing
+            // results in 3 times sending the PING command to redis
+            sleep(2);
+            $db->ping();
+        } catch (SocketException $e) {
+            $exception = true;
+        }
+        $this->assertTrue($exception, 'SocketException should have been thrown.');
+        $this->assertCount(14, $logger->messages, 'log +1 ping command, and reconnection.'
+            . print_r(array_map(function($s) { return (string) $s; }, ArrayHelper::getColumn($logger->messages, 0)), true));
+    }
 
     /**
      * https://github.com/yiisoft/yii2/issues/4745