+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Yii Framework 2 redis extension Change Log
 - Bug #114: Fixed ActiveQuery `not between` and `not` conditions which where not working correctly (cebe, ak1987)
 - Bug #123: Fixed ActiveQuery to work with negative limit values, which are used in ActiveDataProvider for the count query (cebe)
 - Enh #9: Added orderBy support to redis ActiveQuery and LuaScriptBuilder (valinurovam)
+- Enh #91: Added option to retry connection after failing to communicate with redis server on stale socket (cebe)
+- Enh #106: Improved handling of connection errors and introduced `yii\redis\SocketException` for these (cebe)
 - Chg #127: Added PHP 7.2 compatibility (brandonkelly)